azure_function/requirements.txt
azure-functions
pyodbc
python-dotenv
requests
